pyproject.toml
[project]
name = "discord-notifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.5",
    "python-dotenv>=0.21.0",
    "load-dotenv>=0.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.3",
]
